move ExampleMountItem to new Mounts folder
