+++ ExampleMod/Content/Items/Mounts/ExampleMountItem.cs
@@ -1,8 +1,9 @@
+using ExampleMod.Content.Mounts;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace ExampleMod.Content.Items
+namespace ExampleMod.Content.Items.Mounts
 {
 	public class ExampleMountItem : ModItem
 	{
@@ -16,7 +17,7 @@ namespace ExampleMod.Content.Items
 			Item.rare = ItemRarityID.Green;
 			Item.UseSound = SoundID.Item79; // What sound should play when using the item
 			Item.noMelee = true; // this item doesn't do any melee damage
-			Item.mountType = ModContent.MountType<Mounts.ExampleMount>();
+			Item.mountType = ModContent.MountType<ExampleMount>();
 		}
 
 		// Please see Content/ExampleRecipes.cs for a detailed explanation of recipe creation.
