+++ ExampleMod/Content/Items/Mounts/ExampleMountItem.png
